fix: update types
