+++ src/runtime/transformers/index.ts
@@ -1,6 +1,6 @@
 import { extname } from 'pathe'
 import { camelCase } from 'scule'
-import type { ContentTransformer, TransformContentOptions } from './types'
+import type { ContentTransformer, TransformContentOptions } from '../types'
 import csv from './csv'
 import markdown from './markdown'
 import yaml from './yaml'
