+++ src/runtime/transformers/utils.ts
@@ -1,4 +1,4 @@
-import { ContentTransformer } from './types'
+import { ContentTransformer } from '../types'
 
 export const defineTransformer = (transformer: ContentTransformer) => {
   return transformer
