+++ src/runtime/types.d.ts
@@ -135,6 +135,19 @@ export interface MarkdownParsedContent extends ParsedContent {
   }
 }
 
+export interface ContentTransformer {
+  name: string
+  extensions: string[]
+  parse?(id: string, content: string, options: any): Promise<ParsedContent> | ParsedContent
+  transform?(content: ParsedContent, options: any): Promise<ParsedContent> | ParsedContent
+}
+
+export interface TransformContentOptions {
+  transformers?: ContentTransformer[]
+
+  [key: string]: any
+}
+
 /**
  * Query
  */
